(GH-1936, GH-3408) Improvement for title bar and window commands
If there are more then 2 or 3 window commands on the right or left site and the user resizes the window then the Min, Max, Close buttons move out the window and open dialogs will not resize to a slower size.

To solve this I changed the window commands control from ItemsControl to ToolBar class which allows us to use the overlow mechanism of this control.

The default window style is now changed to use this new mechanism.

These changes causes also a breaking change to the window commands overlay behavior. This feature will be removed, because a window command should not overlay when a Flyout is open.

The clean centered style still needs changes for the centered title behavior.